Finished testing build and deploy over lan addon restart

+++ src/Scripts/Build and deploy over LAN.ps1
@@ -4,15 +4,17 @@
 $solutionPath = "D:\source\repos\PockyBum522 Github\netdaemon-home-assistant-apps\src\AllenStreetNetDaemonApps"
 $solutionFilename = "AllenStreetNetDaemonApps.csproj"
 
-$buildPath = "\\192.168.1.25\config\netdaemon4\"
+$buildPath = "\\192.168.1.25\config\netdaemon4"
 
 # Step 1: Run dotnet build
 Set-Location $solutionPath
 
+New-Item -Path $buildPath\apps -ItemType Directory -Force
+
 # Step 2: Delete everything in the network share except 'logs' folder
 Write-Host "Cleaning network share..."
-Get-ChildItem -Path $buildPath -Exclude "logs" | ForEach-Object {
-    if ($_.PSIsContainer -and $_.Name -ne "logs") {
+Get-ChildItem -Path $buildPath -Exclude "logs","apps" | ForEach-Object {
+    if ($_.PSIsContainer -and $_.Name -ne "logs" -and $_.Name -ne "apps") {
         Remove-Item $_.FullName -Recurse -Force -ErrorAction SilentlyContinue
     } elseif (!$_.PSIsContainer) {
         Remove-Item $_.FullName -Force -ErrorAction SilentlyContinue
@@ -24,5 +26,9 @@ Get-ChildItem -Path $buildPath -Exclude "logs" | ForEach-Object {
 
 & dotnet build $solutionPath\$solutionFilename --output $buildPath
 
+Write-Host "Built and deployed, now restarting ND addon"
+
+ssh pockybum522-ha-ssh@192.168.1.25 -i D:\keys\DAVID-LAPTOP-2023-12-24_TO_HOME_ASSISTANT_SSH_ADDON -p 3248 'sudo docker restart addon_c6a2317c_netdaemon4'
+
 # End of the script
-Write-Host "Script execution completed."
+Write-Host "ND addon restarted - Script execution completed."
